[rbrowsable] split item and element for SysFile and TObject
