+++ gui/browsable/src/TDirectoryElement.cxx
@@ -1,20 +1,17 @@
 /*************************************************************************
- * Copyright (C) 1995-2019, Rene Brun and Fons Rademakers.               *
+ * Copyright (C) 1995-2020, Rene Brun and Fons Rademakers.               *
  * All rights reserved.                                                  *
  *                                                                       *
  * For the licensing terms see $ROOTSYS/LICENSE.                         *
  * For the list of contributors see $ROOTSYS/README/CREDITS.             *
  *************************************************************************/
 
-
-#include <ROOT/RBrowsableTDirectory.hxx>
-
 #include <ROOT/Browsable/RLevelIter.hxx>
 #include <ROOT/Browsable/RProvider.hxx>
+#include <ROOT/Browsable/TObjectHolder.hxx>
+#include <ROOT/Browsable/TKeyItem.hxx>
 
-#include <ROOT/RBrowsableTObject.hxx>
-
-#include "ROOT/RLogger.hxx"
+#include <ROOT/RLogger.hxx>
 
 #include "TSystem.h"
 #include "TKey.h"
